MemStatusesテーブルとUsersテーブルで1対多のリレーション構築

+++ cms/database/migrations/2021_06_22_003939_create_gyms_table.php
@@ -22,7 +22,7 @@ class CreateGymsTable extends Migration
             $table->integer('gymType_id');
             $table->integer('superHost_flg');
             $table->integer('review_amount');
-            $table->integer('review_average');
+            $table->float('review_average');
             $table->integer('guest_limit');
             $table->timestamps();
         });
